@@ -3,7 +3,7 @@
 
     <PropertyGroup>
     <OutputType>WinExe</OutputType>
-    <TargetFramework>net7.0-windows</TargetFramework>
+    <TargetFramework>net8.0-windows</TargetFramework>
     <UseWPF>true</UseWPF>
     <RootNamespace>Flora</RootNamespace>
     <EnableDefaultItems>False</EnableDefaultItems>
@@ -13,6 +13,11 @@
   <ItemGroup>
     <PackageReference Include="FontAwesome" Version="4.7.0" />
     <PackageReference Include="MahApps.Metro.IconPacks" Version="4.11.0" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Design" Version="8.0.3">
+      <PrivateAssets>all</PrivateAssets>
+      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
+    </PackageReference>
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="8.0.3" />
     <PackageReference Include="Telerik.UI.for.Wpf.70.Xaml.Trial" Version="2024.1.130" />
   </ItemGroup>
   
@@ -77,12 +82,19 @@
     <DependentUpon>App.xaml</DependentUpon>
     <SubType>Code</SubType>
   </Compile>
+  <Compile Include="DbContexts\MyShopContext.cs" />
   <Compile Include="MainWindow.xaml.cs">
     <DependentUpon>MainWindow.xaml</DependentUpon>
     <SubType>Code</SubType>
   </Compile>
+  <Compile Include="Model\Coupon.cs" />
+  <Compile Include="Model\Customer.cs" />
   <Compile Include="Model\Order.cs" />
-  <Compile Include="Model\Voucher.cs" />
+  <Compile Include="Model\OrderDetail.cs" />
+  <Compile Include="Model\Plant.cs" />
+  <Compile Include="Model\PlantCategory.cs" />
+  <Compile Include="Model\PreviewOrder.cs" />
+  <Compile Include="Model\UserAccount.cs" />
   <Compile Include="Utilities\ButtonCustom.cs" />
   <Compile Include="Utilities\RelayCommand.cs" />
   <Compile Include="Utilities\ViewModelBase.cs" />
@@ -132,6 +144,10 @@
     <Content Include="Images\search-image.png">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </Content>
+  </ItemGroup>
+  
+  <ItemGroup>
+    <None Include="App.config" />
   </ItemGroup>
 
 </Project>
